+++ core/ecosystem/nuget/const.go
@@ -1,0 +1,5 @@
+package nuget
+
+const (
+	EcosystemName = "NuGet"
+)